Añadimos el metodod de alta, edicion y baja de clientes

+++ database/migrations/2022_02_11_000019_create_clients_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('home_address');
-            $table->integer('phone_number');
+            $table->bigInteger('phone_number');
             $table->timestamps();
         });
     }
